+++ package.json
@@ -52,6 +52,7 @@
     "@material/list": "^0.37.0",
     "@material/notched-outline": "^0.35.0",
     "@material/ripple": "^0.37.0",
+    "@material/select": "^0.37.1",
     "@material/textfield": "^0.37.0",
     "@material/top-app-bar": "^0.37.0",
     "@material/typography": "^0.35.0",
