+++ packages/select/index.scss
@@ -1,3 +1,1 @@
-@import "./icon/index.scss";
-@import "./helper-text/index.scss";
-@import "@material/textfield/mdc-text-field";
+@import "@material/select/mdc-select";
