+++ packages/select/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "@material/react-text-field",
-  "version": "0.3.0",
-  "description": "Material Components React Text Field",
+  "name": "@material/react-select",
+  "version": "0.0.0",
+  "description": "Material Components React Select",
   "license": "Apache-2.0",
   "main": "dist/index.js",
   "keywords": [
     "mdc web react",
     "material components react",
     "material design",
-    "text field",
-    "textfield"
+    "select"
   ],
   "repository": {
     "type": "git",
@@ -19,7 +18,7 @@
     "@material/react-floating-label": "^0.3.0",
     "@material/react-line-ripple": "^0.3.0",
     "@material/react-notched-outline": "^0.3.0",
-    "@material/textfield": "^0.37.0",
+    "@material/select": "^0.37.0",
     "classnames": "^2.2.5",
     "prop-types": "^15.6.1",
     "react": "^16.3.2"
