+++ test/screenshot/select/index.scss
@@ -1,7 +1,5 @@
-@import "../../../packages/floating-label/index.scss";
-@import "../../../packages/line-ripple/index.scss";
-@import "../../../packages/notched-outline/index.scss";
+@import "../../../packages/select/index.scss";
 
 body {
-}
   margin: 24px;
+}
